api/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-Cors
gunicorn==20.1.0
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
flask-restx
pyarrow
pandas
sodapy